Response Generator made more fleshed out
I haven't had time to test this yet, but I'm pushing it anyway. The way it works now is the response generator will sort the answers it gets by relevancy score. It will then check if the 2nd most relevant has a relevancy score within 8% (this number was chosen arbitrarily) of the best. If it does, it will add all valid answers to the hold. If not, it will only add the most relevant to the hold. If the best score is under 30 (again, chosen arbitrarily), no answers will be given.